feat: move components folder to src and add barrel
