+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './Content';
+export * from './Footer';
+export * from './Header';
